+++ src/Aws/CloudWatchLogs/Resources/cloudwatchlogs-2014-03-28.php
@@ -330,6 +330,61 @@ return array (
                 ),
             ),
         ),
+        'DeleteSubscriptionFilter' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\JsonCommand',
+            'responseClass' => 'EmptyOutput',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Content-Type' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'application/x-amz-json-1.1',
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/json',
+                ),
+                'X-Amz-Target' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'Logs_20140328.DeleteSubscriptionFilter',
+                ),
+                'logGroupName' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'json',
+                    'minLength' => 1,
+                    'maxLength' => 512,
+                ),
+                'filterName' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'json',
+                    'minLength' => 1,
+                    'maxLength' => 512,
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'Returned if a parameter of the request is incorrectly specified.',
+                    'class' => 'InvalidParameterException',
+                ),
+                array(
+                    'reason' => 'Returned if the specified resource does not exist.',
+                    'class' => 'ResourceNotFoundException',
+                ),
+                array(
+                    'reason' => 'Returned if multiple requests to update the same resource were in conflict.',
+                    'class' => 'OperationAbortedException',
+                ),
+                array(
+                    'reason' => 'Returned if the service cannot complete the request.',
+                    'class' => 'ServiceUnavailableException',
+                ),
+            ),
+        ),
         'DescribeLogGroups' => array(
             'httpMethod' => 'POST',
             'uri' => '/',
@@ -470,6 +525,67 @@ return array (
                     'static' => true,
                     'location' => 'header',
                     'default' => 'Logs_20140328.DescribeMetricFilters',
+                ),
+                'logGroupName' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'json',
+                    'minLength' => 1,
+                    'maxLength' => 512,
+                ),
+                'filterNamePrefix' => array(
+                    'type' => 'string',
+                    'location' => 'json',
+                    'minLength' => 1,
+                    'maxLength' => 512,
+                ),
+                'nextToken' => array(
+                    'type' => 'string',
+                    'location' => 'json',
+                    'minLength' => 1,
+                ),
+                'limit' => array(
+                    'type' => 'numeric',
+                    'location' => 'json',
+                    'minimum' => 1,
+                    'maximum' => 50,
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'Returned if a parameter of the request is incorrectly specified.',
+                    'class' => 'InvalidParameterException',
+                ),
+                array(
+                    'reason' => 'Returned if the specified resource does not exist.',
+                    'class' => 'ResourceNotFoundException',
+                ),
+                array(
+                    'reason' => 'Returned if the service cannot complete the request.',
+                    'class' => 'ServiceUnavailableException',
+                ),
+            ),
+        ),
+        'DescribeSubscriptionFilters' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\JsonCommand',
+            'responseClass' => 'DescribeSubscriptionFiltersResponse',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Content-Type' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'application/x-amz-json-1.1',
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/json',
+                ),
+                'X-Amz-Target' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'Logs_20140328.DescribeSubscriptionFilters',
                 ),
                 'logGroupName' => array(
                     'required' => true,
@@ -905,6 +1021,83 @@ return array (
                 ),
             ),
         ),
+        'PutSubscriptionFilter' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\JsonCommand',
+            'responseClass' => 'EmptyOutput',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Content-Type' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'application/x-amz-json-1.1',
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/json',
+                ),
+                'X-Amz-Target' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'Logs_20140328.PutSubscriptionFilter',
+                ),
+                'logGroupName' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'json',
+                    'minLength' => 1,
+                    'maxLength' => 512,
+                ),
+                'filterName' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'json',
+                    'minLength' => 1,
+                    'maxLength' => 512,
+                ),
+                'filterPattern' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'json',
+                    'maxLength' => 512,
+                ),
+                'destinationArn' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'json',
+                    'minLength' => 1,
+                ),
+                'roleArn' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'json',
+                    'minLength' => 1,
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'Returned if a parameter of the request is incorrectly specified.',
+                    'class' => 'InvalidParameterException',
+                ),
+                array(
+                    'reason' => 'Returned if the specified resource does not exist.',
+                    'class' => 'ResourceNotFoundException',
+                ),
+                array(
+                    'reason' => 'Returned if multiple requests to update the same resource were in conflict.',
+                    'class' => 'OperationAbortedException',
+                ),
+                array(
+                    'reason' => 'Returned if you have reached the maximum number of resources that can be created.',
+                    'class' => 'LimitExceededException',
+                ),
+                array(
+                    'reason' => 'Returned if the service cannot complete the request.',
+                    'class' => 'ServiceUnavailableException',
+                ),
+            ),
+        ),
         'TestMetricFilter' => array(
             'httpMethod' => 'POST',
             'uri' => '/',
@@ -1078,6 +1271,44 @@ return array (
                                         ),
                                     ),
                                 ),
+                            ),
+                            'creationTime' => array(
+                                'type' => 'numeric',
+                            ),
+                        ),
+                    ),
+                ),
+                'nextToken' => array(
+                    'type' => 'string',
+                    'location' => 'json',
+                ),
+            ),
+        ),
+        'DescribeSubscriptionFiltersResponse' => array(
+            'type' => 'object',
+            'additionalProperties' => true,
+            'properties' => array(
+                'subscriptionFilters' => array(
+                    'type' => 'array',
+                    'location' => 'json',
+                    'items' => array(
+                        'name' => 'SubscriptionFilter',
+                        'type' => 'object',
+                        'properties' => array(
+                            'filterName' => array(
+                                'type' => 'string',
+                            ),
+                            'logGroupName' => array(
+                                'type' => 'string',
+                            ),
+                            'filterPattern' => array(
+                                'type' => 'string',
+                            ),
+                            'destinationArn' => array(
+                                'type' => 'string',
+                            ),
+                            'roleArn' => array(
+                                'type' => 'string',
                             ),
                             'creationTime' => array(
                                 'type' => 'numeric',