+++ src/Aws/CognitoIdentity/Resources/cognitoidentity-2014-06-30.php
@@ -67,7 +67,7 @@ return array (
                         'maxLength' => 128,
                         'data' => array(
                             'shape_name' => 'IdentityProviderName',
-                            'key_pattern' => '/[\\w._-]+/',
+                            'key_pattern' => '/[\\w._/-]+/',
                         ),
                     ),
                 ),
@@ -112,6 +112,56 @@ return array (
                 array(
                     'reason' => 'Thrown when the total number of user pools has exceeded a preset limit.',
                     'class' => 'LimitExceededException',
+                ),
+            ),
+        ),
+        'DeleteIdentities' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\JsonCommand',
+            'responseClass' => 'DeleteIdentitiesResponse',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Content-Type' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'application/x-amz-json-1.1',
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/json',
+                ),
+                'X-Amz-Target' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'AWSCognitoIdentityService.DeleteIdentities',
+                ),
+                'IdentityIdsToDelete' => array(
+                    'required' => true,
+                    'type' => 'array',
+                    'location' => 'json',
+                    'minItems' => 1,
+                    'maxItems' => 60,
+                    'items' => array(
+                        'name' => 'IdentityId',
+                        'type' => 'string',
+                        'minLength' => 1,
+                        'maxLength' => 50,
+                    ),
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'Thrown for missing or bad input parameter(s).',
+                    'class' => 'InvalidParameterException',
+                ),
+                array(
+                    'reason' => 'Thrown when a request is throttled.',
+                    'class' => 'TooManyRequestsException',
+                ),
+                array(
+                    'reason' => 'Thrown when the service encounters an error during processing the request.',
+                    'class' => 'InternalErrorException',
                 ),
             ),
         ),
@@ -308,7 +358,7 @@ return array (
                         'maxLength' => 2048,
                         'data' => array(
                             'shape_name' => 'IdentityProviderName',
-                            'key_pattern' => '/[\\w._-]+/',
+                            'key_pattern' => '/[\\w._/-]+/',
                         ),
                     ),
                 ),
@@ -341,6 +391,10 @@ return array (
                 array(
                     'reason' => 'Thrown when the service encounters an error during processing the request.',
                     'class' => 'InternalErrorException',
+                ),
+                array(
+                    'reason' => 'An exception thrown when a dependent service such as Facebook or Twitter is not responding',
+                    'class' => 'ExternalServiceException',
                 ),
             ),
         ),
@@ -387,7 +441,7 @@ return array (
                         'maxLength' => 2048,
                         'data' => array(
                             'shape_name' => 'IdentityProviderName',
-                            'key_pattern' => '/[\\w._-]+/',
+                            'key_pattern' => '/[\\w._/-]+/',
                         ),
                     ),
                 ),
@@ -421,6 +475,10 @@ return array (
                     'reason' => 'Thrown when the total number of user pools has exceeded a preset limit.',
                     'class' => 'LimitExceededException',
                 ),
+                array(
+                    'reason' => 'An exception thrown when a dependent service such as Facebook or Twitter is not responding',
+                    'class' => 'ExternalServiceException',
+                ),
             ),
         ),
         'GetIdentityPoolRoles' => array(
@@ -445,6 +503,7 @@ return array (
                     'default' => 'AWSCognitoIdentityService.GetIdentityPoolRoles',
                 ),
                 'IdentityPoolId' => array(
+                    'required' => true,
                     'type' => 'string',
                     'location' => 'json',
                     'minLength' => 1,
@@ -515,7 +574,7 @@ return array (
                         'maxLength' => 2048,
                         'data' => array(
                             'shape_name' => 'IdentityProviderName',
-                            'key_pattern' => '/[\\w._-]+/',
+                            'key_pattern' => '/[\\w._/-]+/',
                         ),
                     ),
                 ),
@@ -544,6 +603,10 @@ return array (
                 array(
                     'reason' => 'Thrown when the service encounters an error during processing the request.',
                     'class' => 'InternalErrorException',
+                ),
+                array(
+                    'reason' => 'An exception thrown when a dependent service such as Facebook or Twitter is not responding',
+                    'class' => 'ExternalServiceException',
                 ),
             ),
         ),
@@ -591,7 +654,7 @@ return array (
                         'maxLength' => 2048,
                         'data' => array(
                             'shape_name' => 'IdentityProviderName',
-                            'key_pattern' => '/[\\w._-]+/',
+                            'key_pattern' => '/[\\w._/-]+/',
                         ),
                     ),
                 ),
@@ -672,6 +735,11 @@ return array (
                     'type' => 'string',
                     'location' => 'json',
                     'minLength' => 1,
+                ),
+                'HideDisabled' => array(
+                    'type' => 'boolean',
+                    'format' => 'boolean-string',
+                    'location' => 'json',
                 ),
             ),
             'errorResponses' => array(
@@ -1091,7 +1159,7 @@ return array (
                         'maxLength' => 2048,
                         'data' => array(
                             'shape_name' => 'IdentityProviderName',
-                            'key_pattern' => '/[\\w._-]+/',
+                            'key_pattern' => '/[\\w._/-]+/',
                         ),
                     ),
                 ),
@@ -1131,6 +1199,10 @@ return array (
                 array(
                     'reason' => 'Thrown when the service encounters an error during processing the request.',
                     'class' => 'InternalErrorException',
+                ),
+                array(
+                    'reason' => 'An exception thrown when a dependent service such as Facebook or Twitter is not responding',
+                    'class' => 'ExternalServiceException',
                 ),
             ),
         ),
@@ -1184,7 +1256,7 @@ return array (
                         'maxLength' => 128,
                         'data' => array(
                             'shape_name' => 'IdentityProviderName',
-                            'key_pattern' => '/[\\w._-]+/',
+                            'key_pattern' => '/[\\w._/-]+/',
                         ),
                     ),
                 ),
@@ -1267,6 +1339,28 @@ return array (
                     'items' => array(
                         'name' => 'ARNString',
                         'type' => 'string',
+                    ),
+                ),
+            ),
+        ),
+        'DeleteIdentitiesResponse' => array(
+            'type' => 'object',
+            'additionalProperties' => true,
+            'properties' => array(
+                'UnprocessedIdentityIds' => array(
+                    'type' => 'array',
+                    'location' => 'json',
+                    'items' => array(
+                        'name' => 'UnprocessedIdentityId',
+                        'type' => 'object',
+                        'properties' => array(
+                            'IdentityId' => array(
+                                'type' => 'string',
+                            ),
+                            'ErrorCode' => array(
+                                'type' => 'string',
+                            ),
+                        ),
                     ),
                 ),
             ),